refactor: article update actions
