+++ src/actions/cms/articles/create.ts
@@ -6,5 +6,6 @@ import { createArticle } from '@/server/cms/articles/create'
 export async function createArticleAction(name: string | null, config?: {
     categoryId: number,
 }): Promise<ActionReturn<ExpandedArticle>> {
+    //TODO: auth on permission or visibility to categoryId
     return await createArticle(name, config)
 }
